Making corrections after code review
