+++ package.json
@@ -6,6 +6,7 @@
     "start": "vite",
     "build": "tsc && vite build",
     "lint": "eslint src --ext ts,tsx",
+    "lint:fix": "eslint src/ --fix",
     "preview": "vite preview",
     "test": "vitest --passWithNoTests"
   },
@@ -23,11 +24,11 @@
     "react-router-dom": "6.16.0"
   },
   "devDependencies": {
+    "@faker-js/faker": "^8.4.1",
     "@jedmao/redux-mock-store": "3.0.5",
     "@testing-library/jest-dom": "6.1.3",
     "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "14.5.1",
-    "@types/faker": "5.5.9",
     "@types/leaflet": "1.9.3",
     "@types/react": "18.2.25",
     "@types/react-dom": "18.2.11",
@@ -41,7 +42,6 @@
     "eslint-config-htmlacademy": "9.1.1",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-react-refresh": "0.4.3",
-    "faker": "5.5.3",
     "jsdom": "22.1.0",
     "prettier": "3.3.3",
     "typescript": "5.4.2",
