+++ src/components/bookmark/bookmark.tsx
@@ -1,7 +1,7 @@
-import { JSX } from 'react';
+import type { JSX } from 'react';
 import clsx from 'clsx';
 import type { Size } from '@customTypes/size';
-import { getImageSize } from '@functions';
+import { getImageSize } from '@utils';
 
 type PrefixForBookmark = 'place-card' | 'offer';
 
@@ -12,11 +12,15 @@ type BookmarkProps = {
 };
 
 function Bookmark({ size, prefix, isFavorite }: BookmarkProps): JSX.Element {
+  const bookmarkLabel = `${isFavorite ? 'In' : 'To'} bookmarks`;
+
   return (
     <button
-      className={clsx(`${prefix}__bookmark-button`, 'button', {
-        [`${prefix}__bookmark-button--active`]: isFavorite,
-      })}
+      className={clsx(
+        `${prefix}__bookmark-button`,
+        'button',
+        isFavorite && `${prefix}__bookmark-button--active`,
+      )}
       type="button"
     >
       <svg
@@ -25,7 +29,7 @@ function Bookmark({ size, prefix, isFavorite }: BookmarkProps): JSX.Element {
       >
         <use xlinkHref="#icon-bookmark" />
       </svg>
-      <span className="visually-hidden">To bookmarks</span>
+      <span className="visually-hidden">{bookmarkLabel}</span>
     </button>
   );
 }
