+++ src/components/favorites-offers/index.tsx
@@ -1,0 +1,1 @@
+export { FavoritesOffers } from './favorites-offers';
