+++ src/components/form-rating-star/index.tsx
@@ -1,0 +1,1 @@
+export { FormRatingStar } from './form-rating-star';
