+++ src/components/form-rating/index.tsx
@@ -1,0 +1,1 @@
+export { FormRating } from './form-rating.tsx';
