+++ src/components/location-item/location-item.tsx
@@ -1,7 +1,7 @@
-import { JSX } from 'react';
+import type { JSX } from 'react';
 import type { CityName } from '@customTypes/city';
 
-type LocationItemTags = 'li' | 'div';
+type LocationItemTags = keyof Pick<JSX.IntrinsicElements, 'li' | 'div'>;
 
 type LocationItemProps = {
   city: CityName;
@@ -11,17 +11,17 @@ type LocationItemProps = {
 
 function LocationItem({
   city,
-  containerTag,
+  containerTag: TagName,
   linkClassName,
 }: LocationItemProps): JSX.Element {
-  const Container: keyof JSX.IntrinsicElements = containerTag;
+  // const Container: keyof JSX.IntrinsicElements = containerTag;
 
   return (
-    <Container className="locations__item">
+    <TagName className="locations__item">
       <a className={linkClassName} href="#">
         <span>{city}</span>
       </a>
-    </Container>
+    </TagName>
   );
 }
