+++ src/components/locations-list/index.tsx
@@ -1,0 +1,1 @@
+export { LocationsList } from './locations-list';
