+++ src/components/login-form/login-form.tsx
@@ -1,4 +1,4 @@
-import { JSX } from 'react';
+import type { JSX } from 'react';
 
 function LoginForm(): JSX.Element {
   return (
