+++ src/components/logo/logo.tsx
@@ -1,6 +1,6 @@
-import { JSX } from 'react';
+import type { JSX } from 'react';
 import clsx from 'clsx';
-import { getImageSize } from '@functions';
+import { getImageSize } from '@utils';
 import type { Size } from '@customTypes/size';
 
 type PrefixForLogo = 'header' | 'footer';
@@ -13,9 +13,10 @@ type LogoProps = {
 function Logo({ size, prefix, isActive = false }: LogoProps): JSX.Element {
   return (
     <a
-      className={clsx(`${prefix}__logo-link`, {
-        [`${prefix}__logo-link--active`]: isActive,
-      })}
+      className={clsx(
+        `${prefix}__logo-link`,
+        isActive && `${prefix}__logo-link--active`,
+      )}
     >
       <img
         className={`${prefix}__logo`}
