+++ src/components/main-offers/index.tsx
@@ -1,0 +1,1 @@
+export { MainOffers } from './main-offers';
