+++ src/components/main-offers/main-offers.tsx
@@ -1,0 +1,56 @@
+import type { JSX } from 'react';
+import { OFFERS } from '@mocks/offers';
+import { OfferCard } from '@components/offer-card';
+
+function MainOffers(): JSX.Element {
+  const countOffers: number = OFFERS.length;
+
+  return (
+    <>
+      <section className="cities__places places">
+        <h2 className="visually-hidden">Places</h2>
+        <b className="places__found">
+          {countOffers} places to stay in Amsterdam
+        </b>
+        <form className="places__sorting" action="#" method="get">
+          <span className="places__sorting-caption">Sort by</span>
+          <span className="places__sorting-type" tabIndex={0}>
+            Popular
+            <svg className="places__sorting-arrow" width="7" height="4">
+              <use xlinkHref="#icon-arrow-select" />
+            </svg>
+          </span>
+          <ul className="places__options places__options--custom places__options--opened">
+            <li className="places__option places__option--active" tabIndex={0}>
+              Popular
+            </li>
+            <li className="places__option" tabIndex={0}>
+              Price: low to high
+            </li>
+            <li className="places__option" tabIndex={0}>
+              Price: high to low
+            </li>
+            <li className="places__option" tabIndex={0}>
+              Top rated first
+            </li>
+          </ul>
+        </form>
+        <div className="cities__places-list places__list tabs__content">
+          {OFFERS.map((offer) => (
+            <OfferCard
+              key={offer.id}
+              offer={offer}
+              size="large"
+              prefix="cities"
+            />
+          ))}
+        </div>
+      </section>
+      <div className="cities__right-section">
+        <section className="cities__map map"></section>
+      </div>
+    </>
+  );
+}
+
+export { MainOffers };
